#391 edit label, tooltip condition in combine chart
