+++ discovery-frontend/src/app/common/component/chart/option/converter/format-option-converter.ts
@@ -794,9 +794,10 @@ export class FormatOptionConverter {
 
       case ChartType.BAR:
       case ChartType.LINE:
+      case ChartType.COMBINE:
         // when bar, line chart has single series
         if ((chartType === ChartType.BAR && pivot.aggregations.length <= 1 && pivot.rows.length < 1) ||
-          (chartType === ChartType.LINE && pivot.aggregations.length <= 1)) {
+          ((chartType === ChartType.LINE || chartType === ChartType.COMBINE) && pivot.aggregations.length <= 1)) {
           displayTypes[0] = UIChartDataLabelDisplayType.CATEGORY_NAME;
           displayTypes[1] = UIChartDataLabelDisplayType.CATEGORY_VALUE;
           // when bar, line chart has multi series
@@ -807,7 +808,6 @@ export class FormatOptionConverter {
         break;
 
       case ChartType.CONTROL:
-      case ChartType.COMBINE:
       case ChartType.WATERFALL:
         displayTypes[0] = UIChartDataLabelDisplayType.CATEGORY_NAME;
         displayTypes[1] = UIChartDataLabelDisplayType.CATEGORY_VALUE;
